Htools should use state-of-record instance size
The actual memory usage is irrelevant, because the instance can decide anytime
to touch the yet unused portion, so we can't take away that memory from
the instance without risking OOM conditions.

In case of oversubscription (unsupported yet) we have the following mechanisms:

KVM: Memory manager has to evict some pages to swap. That means:
 swap + physical_memory > sum(all instances state-of-record size) + node_memory

Xen: Balooning has to ask the instance's permission to take some pages by
allocating a buffer in the instance's balooning kernel driver. Thus the instance
isn't in fact resized, just some pages are borrowed and reused in the free pool.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>