Add optStaticNodeMemory to hbal
This is an attempt to fix balancing on 2.16 using KVM hypervisor. The problem
with KVM is, that it reports 'active' memory from /proc/memstat for node
memory a.k.a. nMem a.k.a. memory_dom0.

This is plain wrong, as 'active' memory in Linux includes memory from all
running processes, including instances (qemu processes). We can't calculate
the node memory on the hypervisor side, simply subtracting RES used by qemu
processes would give us only approximate value and still it would be the actively
used memory rather than the maximum reserved as compared to dom0 in Xen. As KVM
does not provide proper isolation for node memory, the only thing we can do is
to instruct hbal to reserve a static value for node memory.

This is solved in 2.17 in obtainNodeMemory function that takes the live value
from Xen, but uses a configurable cluster parameter for KVM.

I've opted for not backporting the cluster parameter from 2.17 for two reasons:
* this way the change is isolated to hbal, no need to change RPC's and data
* the goal is to fix balancing without memory overcommitment, that was a feature
intended for 2.17

Signed-off-by: Viktor Bachraty <vbachraty@google.com>